+++ tests/SharedTestProjectProperties.xml
@@ -7,5 +7,8 @@
 		<Compile Include="..\QueryableValues.EF6.SqlServer.Tests.EF60\DatabaseFirst\**\*.cs">
 			<Link>DatabaseFirst\%(RecursiveDir)\%(FileName)%(Extension)</Link>
 		</Compile>
+		<Compile Include="..\QueryableValues.EF6.SqlServer.Tests.EF60\Queries\**\*.cs">
+			<Link>Queries\%(RecursiveDir)\%(FileName)%(Extension)</Link>
+		</Compile>
 	</ItemGroup>
 </Project>
